requirements.txt
algoliasearch==2.6.2
whisper==1.1.10
altair==4.2.0
appdirs==1.4.4
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
DateTime==4.5
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
exabel-data-sdk==4.0.0
finnhub-python==2.4.13
Flask==2.1.3
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-play-scraper==1.2.2
googleapis-common-protos==1.56.4
grpcio==1.50.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lxml==4.9.1
MarkupSafe==2.1.1
multitasking==0.0.11
numpy==1.23.4
openpyxl==3.0.10
packaging==22.0
pandas==1.4.3
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
streamlit==1.16.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
Werkzeug==2.2.2
xgboost==1.7.2
yfinance==0.1.74
zipp==3.11.0
zope.interface==5.5.1
deepgram-sdk==2.3.0
nltk==3.8.1